Implement nested wrappers
In the case WrapExpr (... WrapExpr (... HsVar)), with the var modified,
we now find the wrapper which binds the last TyVar occurring in the type
of the var, and add the new EvApps to the outside of that wrapper. This
works because instantiateSigma always puts a set of TyApps and the
following EvApps in the same wrapper (I think), and our new constraints
always go at the end of the invis binders.